requirements.txt
pygame == 1.9.6
pillow==5.4.1